add forge build command

+++ include/forge/core/Version.hpp
@@ -1,16 +1,15 @@
 #ifndef FORGE_VERSION_HPP
 #define FORGE_VERSION_HPP
 
-#include <cstdint>
 #include <string>
 
 namespace Forge::Core
 {
     struct Version
     {
-        uint8_t major{};
-        uint8_t minor{};
-        uint8_t patch{};
+        unsigned int major{};
+        unsigned int minor{};
+        unsigned int patch{};
         Version() = default;
         explicit Version(const std::string& versionStr);
         [[nodiscard]] auto to_tuple() const noexcept;
